refactor: apply linter
