+++ src/components/userCard/UserCard.module.scss
@@ -9,7 +9,6 @@
   text-align: center;
   height: 350px;
 
-
   &:hover {
     transform: scale(1.01);
   }
